Add Airflow database backup DAG
- Introduced a new DAG `airflow_db_backup` for backing up the Airflow Postgres database.
- The DAG compresses the database dump and uploads it to S3 with a date-based key.
- Added tasks to load AWS credentials from 1Password and include the current date in the upload path.
- Configured the DAG to run daily at 3:00 AM in production environment.
